requirements.txt
djangorestframework==3.14.0
pip-chill==1.0.3
django==4.2.2